further DataSourceTests cleanup
