+++ Tests.MemoQ/DataSourceTests.cs
@@ -1,9 +1,7 @@
-﻿using Apps.Memoq.Actions;
+﻿using Tests.MemoQ.Base;
 using Apps.Memoq.DataSourceHandlers;
-using Apps.Memoq.Models.ServerProjects.Requests;
 using Apps.MemoQ.DataSourceHandlers;
 using Blackbird.Applications.Sdk.Common.Dynamic;
-using Tests.MemoQ.Base;
 
 namespace Tests.MemoQ;
 
@@ -21,51 +19,6 @@ public class DataSourceTests : TestBase
             Console.WriteLine($"{item.Value}: {item.DisplayName}");
         }
         Assert.IsTrue(result.Count() > 0);
-    }
-
-    [TestMethod]
-    public async Task Projects_returns_report_values()
-    {
-        var handler = new ServerProjectActions(InvocationContext, FileManager);
-
-        var result = await handler.SearchEditReports(new Apps.Memoq.Models.ServerProjects.Requests.ProjectRequest { ProjectGuid = "10bd767d-3ce2-ef11-875f-a8a15994f72e" });
-        foreach (var item in result.Reports)
-        {
-            Console.WriteLine($"{item.ReportId}: {item.Note}");
-        }
-        Assert.IsNotNull(result);   
-    }
-
-    [TestMethod]
-    public async Task Report_returns_report_values()
-    {
-        var handler = new FileActions(InvocationContext, FileManager);
-
-        var result = await handler.GetEditDistanceReport(
-            new Apps.Memoq.Models.ServerProjects.Requests.ProjectRequest { ProjectGuid = "10bd767d-3ce2-ef11-875f-a8a15994f72e" },
-            "a25715b6-c004-f011-875f-a8a15994f72e");
-       
-        Assert.IsNotNull(result);
-    }  
-
-    [TestMethod]
-    public async Task UpdateProject_returns_values()
-    {
-        var handler = new ServerProjectActions(InvocationContext, FileManager);
-
-        await handler.UpdateProject(new ProjectRequest
-        {
-            ProjectGuid = "3c8014cf-c3f0-ed11-85f6-d05099f919f4"
-        }, new UpdateProjectRequest 
-        { 
-            CallbackUrl= "https://webhook.test.com", 
-            Description= "Test description 2",
-            Subject="Test subject 3",
-            Domain= "Test domain 2",
-            Client= "Test client 2",
-            Deadline = DateTime.Now.AddDays(7)
-        });
-        Assert.IsTrue(true);
     }
 
     [TestMethod]
