+++ Tests.MemoQ/ServerProjectTests.cs
@@ -10,6 +10,38 @@ namespace Tests.MemoQ;
 public class ServerProjectTests : TestBase
 {
     [TestMethod]
+    public async Task Projects_returns_report_values()
+    {
+        var handler = new ServerProjectActions(InvocationContext, FileManager);
+
+        var result = await handler.SearchEditReports(new ProjectRequest { ProjectGuid = "10bd767d-3ce2-ef11-875f-a8a15994f72e" });
+        foreach (var item in result.Reports)
+        {
+            Console.WriteLine($"{item.ReportId}: {item.Note}");
+        }
+        Assert.IsNotNull(result);
+    }
+
+    [TestMethod]
+    public async Task UpdateProject_returns_values()
+    {
+        var handler = new ServerProjectActions(InvocationContext, FileManager);
+
+        await handler.UpdateProject(new ProjectRequest
+        {
+            ProjectGuid = "3c8014cf-c3f0-ed11-85f6-d05099f919f4"
+        }, new UpdateProjectRequest
+        {
+            CallbackUrl = "https://webhook.test.com",
+            Description = "Test description 2",
+            Subject = "Test subject 3",
+            Domain = "Test domain 2",
+            Client = "Test client 2",
+            Deadline = DateTime.Now.AddDays(7)
+        });
+    }
+
+    [TestMethod]
     public async Task SetCustomField_IsSuccess()
     {
         var actions = new ServerProjectActions(InvocationContext, FileManager);
